Load from image feature moved to a dedicated folder
