+++ src/components/LoadFromFile/ImageSizeDisplay.jsx
